HADOOP-5838. Fix a few javac warnings in HDFS. (Raghu Angadi)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@775210 13f79535-47bb-0310-9956-ffa450edef68

+++ src/hdfs/org/apache/hadoop/hdfs/server/balancer/Balancer.java
@@ -641,7 +641,7 @@ public class Balancer implements Tool {
      */
     private long getBlockList() throws IOException {
       BlockWithLocations[] newBlocks = namenode.getBlocks(datanode, 
-        (long)Math.min(MAX_BLOCKS_SIZE_TO_FETCH, blocksToReceive)).getBlocks();
+        Math.min(MAX_BLOCKS_SIZE_TO_FETCH, blocksToReceive)).getBlocks();
       long bytesReceived = 0;
       for (BlockWithLocations blk : newBlocks) {
         bytesReceived += blk.getBlock().getNumBytes();
